Add script to test rubyfmt against any Git repo.
This commit adds a script to run all of the Ruby files in some external Git
repo through rubyfmt, only outputting errors and the paths to the files
where they were encountered.

The purpose of the script is to make it easier to throw a whole bunch of
code at rubyfmt, so that we can more easily find bugs and uncovered
edge-cases.