Normalisation experiment (#117)
* Only run rubyfmt when not required.

This commit makes it possible to `require "rubyfmt"` in another script, e.g.
to quickly test some experimental feature.

* Normalize method calls before formatting.

What?
=====

Ripper has a _lot_ of ways it can represent method calls. They include:

- `method_add_arg`, for any call with parens
- `command`, for a bare method call without parens
- `command_call`, for dotted method call without parens
- `call`, for a dotted method call without parens or arguments
- `vcall`, for a bare method call without parens or arguments

There are also some very method-call-like things, where the same formatting
rules probably ought to apply. They are:

- `zsuper`, for super with no arguments
- `super`, for super with arguments, or with empty parens
- `yield`
- `return`

This commit normalises these s-expressions into the following structure
before formatting them:

    [
      :method_call,
      receiving_object,
      method_name,
      original_used_parens,
      arguments,
    ]

Where:

- `receiving_object` is an array of expressions, which can be concatenated
  to produce the receiving object. This could include operators like `.`,
  `&.`, or `::`.

- `method_name` is an expression representing the name of the method.

- `original_used_parens` is a boolean, indicating if the original source
  code used parens. This is useful for `super` calls, where empty parens
  change the semantics.

- `arguments` is an array or expressions representing the arguments.

To simplify further, a few other s-expressions are also modified. In the
`receiving_object` array:

- Operators are represented as `[:dot, expression_or_symbol]`. This fits
  neatly with the existing `#format_dot` method.

And in the `arguments` array:

- Splats are represented as `[:spat, expression]` where `expression` is
  the splatted argument.

- Blocks passed as params are represented as `[:to_proc, expression]`
  where `expression` is the block argument.

Now that these cases are all handled consistently, some fixtures have
necessarily changed to make sure they're consistent with each other.

Why?
====

When a subtle change to the formatting of a method call (e.g. the addition
or parentheses) could change the s-expression representation of the
expression (e.g. from a `command` to a `method_add_arg`), and therefore
change the code path followed by rubyfmt, it was easy for differences
between formatting to creep in

There are a couple of reasons this matters:

1. We want rubyfmt to be idempotent, i.e. running a file through it many
   times should produce the same output as running it through once.
   Maintaining that property while formatting the same kind of thing in
   multiple places would be difficult.

2. We're likely to go back-and-forth on what the output should look like.
   This means a single place to make changes to method calls and other
   call-like structures will be beneficial.

* Extract method to decide if we need parens.

* Be more consistent with previous behaviour.

* Use parens with `require` if the source did.

This should fix a failure in the RSpec stress test, but maybe isn't the
formatting decision we want to make long-term.

The real thing we want here is not to change the semantics of a thing like
`require(path_for(formatter_ref)) ? retry : raise` by dropping some parens.

* Add a stress test.

* fix variable names in test script