Use jemalloc instead of the system allocator (#160)
* Use jemalloc instead of the system allocator

A significant portion of time is spent allocating the `Vec`s in various
rubyfmt data structures. By switching to jemalloc, that cost is
effectively removed, reducing the time to parse a 2kloc file from 140ms
to 105ms on my machine (roughly a 33% improvement)

* Disable "initial exec tls" for jemalloc

This causes issues on CI, and it seems to run even faster without it.