Operate directly on the ripper output without JSON parsing
This removes the json serialization/parsing in favor of linking libruby
directly, and inspecting the ruby objects directly. This is done by just
implementing a minimal `serde::Deserializer` that traverses the Ruby
objects. While this does *not* speed up the creation of the AST, it does
remove the call to `.to_json`, which could be as much as 15% of the
total runtime.

I've used rutie to access libruby. While we aren't using the library
itself (which is a full featured "Rust in Ruby" wrapper), its rubysys
module seems to be the most up to date and maintained bindings to
libcruby.

The biggest issue with accessing ruby's C API from Rust is macros like
`RSTRING_PTR` and `RSTRING_LEN`. Rutie re-implements these macros in
Rust, relying on the internal representation of RBasic. While this is
private, it has also been stable for over a decade now (and a lot of
other code would break if this ever changed). libcruby-sys (which is
part of Helix) goes the route of wrapping these macros in a function
which it exposes, but that requires the helix-runtime gem in the
consuming code. We *could* do this ourselves, but that will require
futzing with linker flags to get this to link against functions that
don't exist at link time (and if we're doing that we may as well just
drop rutie as a dependency and write the `extern "C" blocks ourselves`)

In addition, while I was debugging this impl, I wanted to see the full
error message. I've changed this to raise the actual error instead of
returning an opaque status code (for some reason rutie doesn't expose
`rb_eRuntimeError` so I've just passed that in at the top level).

I've also disabled GC for an additional small win. The end result of all
this is that on my machine, formatting
rspec-core/lib/rspec/core/configuration.rb goes from ~98ms to ~80ms. At
this point any additional optimization will hit serious diminishing
returns. We're spending less than 40% our time in Rust, the majority is
spent in Ruby setup/teardown, `require`, and the actual parsing.

That said, of the time spent in our code, 60% of that is in the
deserialization. Some of that might just be the inherent cost of
constructing all these data structures, but I suspect that serde is
generating less than optimal code for `#[serde(untagged)]`, so I plan on
digging into this further. The other big win could come from not freeing
all the memory we allocate for the AST. However, since it's passed by
value everywhere instead of by reference, this is non-trivial. I suspect
that if we pass by reference everywhere we might see a pretty decent win
anyway, so I'll explore that as well.