Remove rutie, expose ruby.h manually
If there is a hard part of using Rust from Ruby, it's that many of
Ruby's important functions (like getting the pointer for a string) are
exposed as macros or inline functions, rather than symbols which can be
linked against.

Of things that are effectively a "ruby-sys" crate, there are two main
options. rutie's rubysys module, or libcruby-sys, which is part of
Helix.

libcruby-sys takes the same approach as this commit, wrapping anything
that is a macro in a small C function compiled as part of a gem. The
problem is that libcruby-sys then requires that gem as a dependency of
the consumer (helix_runtime). Since we're disabling gems, that's a
non-starter. This function wrapping does technically have some small
runtime cost (if it doesn't get inlined by LTO. I'm not familiar enough
with how C compilation for Ruby gems works to know if LTO is happening
or not), but the difference was not measurable on my machine.

Rutie avoids this by re-implementing those macros in Rust. By definition
anything in those macros has to be in `ruby.h`, so it's technically
public, but it's unclear how stable any of that is. For example, getting
the length of a string involves checking if it's embedded (a.k.a the
short string optimization). While the things that it's coupled to seem
to have historically been relatively stable, it makes me uncomfortable
enough to justify what we're doing.

Ultimately I opted to just write `ruby.rs` manually instead of using
something like bindgen, since we're only using a small handful of
functions from libruby.

I had to set up some linker args to allow unknown symbols in the binary.
While we could use RUBYINFO to get linker args for any symbols which
come from Ruby, there's absolutely no way for us to get the macros which
we're wrapping in a C function at link time, since the Rust needs to be
compiled before the C.