Operate directly on the ripper output without JSON parsing (#161)
* Operate directly on the ripper output without JSON parsing

This removes the json serialization/parsing in favor of linking libruby
directly, and inspecting the ruby objects directly. This is done by just
implementing a minimal `serde::Deserializer` that traverses the Ruby
objects. While this does *not* speed up the creation of the AST, it does
remove the call to `.to_json`, which could be as much as 15% of the
total runtime.

I've used rutie to access libruby. While we aren't using the library
itself (which is a full featured "Rust in Ruby" wrapper), its rubysys
module seems to be the most up to date and maintained bindings to
libcruby.

The biggest issue with accessing ruby's C API from Rust is macros like
`RSTRING_PTR` and `RSTRING_LEN`. Rutie re-implements these macros in
Rust, relying on the internal representation of RBasic. While this is
private, it has also been stable for over a decade now (and a lot of
other code would break if this ever changed). libcruby-sys (which is
part of Helix) goes the route of wrapping these macros in a function
which it exposes, but that requires the helix-runtime gem in the
consuming code. We *could* do this ourselves, but that will require
futzing with linker flags to get this to link against functions that
don't exist at link time (and if we're doing that we may as well just
drop rutie as a dependency and write the `extern "C" blocks ourselves`)

In addition, while I was debugging this impl, I wanted to see the full
error message. I've changed this to raise the actual error instead of
returning an opaque status code (for some reason rutie doesn't expose
`rb_eRuntimeError` so I've just passed that in at the top level).

I've also disabled GC for an additional small win. The end result of all
this is that on my machine, formatting
rspec-core/lib/rspec/core/configuration.rb goes from ~98ms to ~80ms. At
this point any additional optimization will hit serious diminishing
returns. We're spending less than 40% our time in Rust, the majority is
spent in Ruby setup/teardown, `require`, and the actual parsing.

That said, of the time spent in our code, 60% of that is in the
deserialization. Some of that might just be the inherent cost of
constructing all these data structures, but I suspect that serde is
generating less than optimal code for `#[serde(untagged)]`, so I plan on
digging into this further. The other big win could come from not freeing
all the memory we allocate for the AST. However, since it's passed by
value everywhere instead of by reference, this is non-trivial. I suspect
that if we pass by reference everywhere we might see a pretty decent win
anyway, so I'll explore that as well.

* Remove rutie, expose ruby.h manually

If there is a hard part of using Rust from Ruby, it's that many of
Ruby's important functions (like getting the pointer for a string) are
exposed as macros or inline functions, rather than symbols which can be
linked against.

Of things that are effectively a "ruby-sys" crate, there are two main
options. rutie's rubysys module, or libcruby-sys, which is part of
Helix.

libcruby-sys takes the same approach as this commit, wrapping anything
that is a macro in a small C function compiled as part of a gem. The
problem is that libcruby-sys then requires that gem as a dependency of
the consumer (helix_runtime). Since we're disabling gems, that's a
non-starter. This function wrapping does technically have some small
runtime cost (if it doesn't get inlined by LTO. I'm not familiar enough
with how C compilation for Ruby gems works to know if LTO is happening
or not), but the difference was not measurable on my machine.

Rutie avoids this by re-implementing those macros in Rust. By definition
anything in those macros has to be in `ruby.h`, so it's technically
public, but it's unclear how stable any of that is. For example, getting
the length of a string involves checking if it's embedded (a.k.a the
short string optimization). While the things that it's coupled to seem
to have historically been relatively stable, it makes me uncomfortable
enough to justify what we're doing.

Ultimately I opted to just write `ruby.rs` manually instead of using
something like bindgen, since we're only using a small handful of
functions from libruby.

I had to set up some linker args to allow unknown symbols in the binary.
While we could use RUBYINFO to get linker args for any symbols which
come from Ruby, there's absolutely no way for us to get the macros which
we're wrapping in a C function at link time, since the Rust needs to be
compiled before the C.

* Cleanup imports and cargo.toml

Co-authored-by: Penelope Phippen <penelopedotzone@gmail.com>