Re-enable warnings in ripper_tree_types.rs, fix them
This removes the `#![allow(warnings)]` from that file, and fixes (or
allows) all the warnings that occurred. The most notable were several
cases of `clippy::large_enum_variant`, which triggers when an enum
variant is more than 200 bytes larger than the next largest. This
triggered quite a bit because Expression was 864 bytes! Note that I
generally didn't follow the advice of the lint, and generally opted to
make the structs that were the cause smaller. Expression is now 184
bytes. This was a marginal improvement in CPU time, and a significant
improvement in memory usage.