Collapse rendering to a single target stack
This finalizes collapsing rendering to a single stack of rendering
targets, harmonizing the "render queue", which is just a wrapper around
a Vec of tokens, and a BreakableEntry, which has some special state
management for dealing with line splitting of things like arrays,
hashes, method calls, and method definitions.

This moves all the behaviour of knowing which target we're touching out
of the ParserState and in to the new RenderTargetStack struct.

Things I like about this:

* ParserState is smaller, and it's kind of our god object
* We're operating on fewer primitives, and more structs with named
  methods

Things I don't like about this:
* RenderTargetStack knows a lot about whether or not it's currently
  targetting a breakable entry, but maybe that's fine
* We have to expose pushing and popping a BreakableEntry to the
  ParserState, it'd be nice if target_stack could have a
  `with_breakable_entry` function, instead of exposing direct push and
  pop methods, but I couldn't work out how to make the borrow checker
  play nice with it

There is a very minor behaviour change in this commit in that if
BreakableEntries contain hard newlines they will flush comments, but
this didn't break any of the fixtures (largely because we don't have
fixtures for that behaviour). I'm going to fix that up in a future
commit.