Add script for generating a shim rubyfmt script
The README currently recommends setting up a `rubyfmt` alias as part of
the installation process, but depending on how you are invoking the
command, not every environment will work with an alias.  This script
will install a shim script at the location of the user's choice:

``` bash
$ ./script/install_shim.sh ~/bin/rubyfmt
```

The script adds the `RUBYFMT_USE_RELEASE` variable as well so as to
prevent unnecessary debug output.