Failing test for multiline comment in empty method
This test demonstrates a formatting error when there are multiline
comments inside an empty method. When formatting, the first line of the
comment remains in the method, while all other lines get pulled out to
just after the method. Additionally, if the method is the last method in
the file, only the second line gets pulled after `end`. Everything else
in the comment, as well as a comment following the method, gets dropped
entirely.