Fix balanced string delims around interpolation (#212)
Fixes: #211

The previous way of doing this with eval was very clever, but resulted
in trying to eval strings like `%{{}` which is just doesn't work. I
tried to think of ways we could continue to use this eval pattern
somehow but everything i could think of was more fragile than regular
expressions, so we lets use regular expressions.

Because (I assume, please don't tell me this is incorrect) single quoted
and %q strings don't have interpolations I continue to use the original
eval pattern (i tried processing these with regular expressions as well
and they have completely different handling of escapes, so since this
works and is less fragile it can stay.

So we find the characters we need to escape after counting the
backslashes and either insert or remove a backslash as necessary.

Sorry i had to ruin strings again.