Fix unescaped square brackets in percent arrays
e.g. `%w{[ ]}` => `%w[\[ \]]` (which is equivalent to `["[", "]"]`)

fixes: #225

This does mean that this is possibly unnecessarily escaped, because
balanced pairs are fine in percent literals BUT this is easier, and
slightly more explicit.

i can't tell anywhere in the ripper output what kind of delimiters were
used. so this is still wrong: `%w{\{}`: before it would be
equivalent to `['{']`,
now it's transformed to `%w[\{]` which is equivalent to `['\\{']`
but that's a preexisting issue, at least we solved the unbalanced [ case
(because %W/I escapes differently than %w/i, this doesn't appear to be
an issue with those)

(also i'm unclear which ruby file i'm supposed to be modifying so i
modified them all)