Fix double free in `debug_inspect` (#231)
This was the cause of the segfault that earlier commits in this branch
tried to fix. They actually fixed it by not calling `debug_inspect`, the
way it was performing iteration was fine (other than the signature of
the callback missing an argument, which could fuck things up depending
on the ABI of the target).

The source of this bug is that `CString` is being constructed with a
pointer we don't own. `CString` is an owned string, which frees its
memory on drop. This results in a double free at best.

The reason this manifested as it did was because we're not just freeing
a pointer we don't own, we're freeing a pointer that was allocated by a
different allocator. Since we didn't build ruby with jemalloc, and we
aren't compiling our jemalloc with unprefixed symbols (we should totally
do that), jemalloc would end up corrupting its own book keeping when we
freed it.

This is also why the segfault disappeared if we used the system
allocator in Rust, or unprefixed symbols in jemalloc. When we do either
of those, it becomes "just" a double free, since the string is never
read again. Even the double free was unlikely to manifest since GC
almost certainly never runs again after Ripper finishes parsing.