Partial audit of unsafe code, refactor to reduce its usage
This is the result of a partial audit of all unsafe code in rubyfmt. I
found one double free, and many instances where unsafe was used more
than it was needed. In some cases this was caused by omitting extremely
cheap safety checks like UTF-8 validation, and in some cases it was
redundant code converting arrays and strings to Rust versions. I've
added UTF-8 checking where it's clearly not going to cause a performance
problem, commented where it's not clear, and moved array and string
conversion to a common function that's easier to audit.