Implement support for BEGIN and END block parsing
Initial fix for #236.

I believe this is the minimum necessary to support BEGIN / END block
parsing and formatting.

Before this commit the ruby keywords 'BEGIN' and 'END' would cause
rubyfmt to stop and throw an error. This adds support for processing
BEGIN and END and then hands off the rest of formatting to something
that is similar to a when or else block.

This the output should look like:

    BEGIN {
       # rest of program indented approriately within the
       # BEGIN / END block
    }

This may not be the way that rubyfmt wants to format BEGIN and END
blocks. In any case, this skeleton can be the basis for not blowing
up.