Fix bug with yield arguments within ifop
A call to yield with arguments within an IfOp (which is what the ruby
parser calls a ternery hook) would remove the parentheses from the yield
call, which would then produce code that Ruby could not parse because of
the precedence of the ? and : operators.

This commit fixes that by ensuring that within an IfOp context we always
keep the parentheses around a yield call if it has arguments.