Fixes rendering heredocs in method calls.
The problem was that essentially they would always be placed in the
wrong place due to how we assumed they work with newlines. This puts the
heredocs on collapsing newlines if we go to render an abstract target
(like a method call) and then pops them in the right place at the render
queue writer stage.

Fixes #209