Automatically inform the user if a new version is available
Penelope asked me to "build an auto-update checking mechanism that
doesn't slow it down". This should meet that criteria, though not
slowing things down definitely made it trickier.

Once this PR is merged, the version field in `Cargo.toml` will need to
start getting updated whenever a new release is put out. I've also
updated the version field for librubyfmt, but it's not used by the code
in this commit.

To determine what the most recent version is, we hit GitHub's API to get
the tags on the repo, and look for the highest tag that looks like a
version (using semver sorting, not lexographic sorting which is what
GitHub would give us). To avoid spamming their API, we perform this
check at most once per day.

The difficult thing about checking for updates this way is that rubyfmt
is very likely to finish doing its thing faster than we can make that
network request, especially if the user is on a high latency connection.

Because of this, we spawn a detached child process to make the API
request. The version it finds is written to a file, and the mtime of
that file is used to track when the last time we asked GitHub was. The
only work done by the main rubyfmt process is reading that file.

As written, this commit does technically slow it down since I'm blocking
on the file I/O. I believe the impact here should be negligible, though
it might be measurable if rubyfmt is reading from stdin/out instead of
doing file I/O elsewhere. If this is a concern, I can move the file IO
to another thread and have it race, so if we can't read the file to
check the version before formatting finishes, we just exit. I think the
gains from doing this would be marginal at best.

Since we need to have persistent state across runs, the question arises
of where we should actually put this file. I opted to put this in a data
dir, rather than a cache dir, since we don't actually want this file
getting cleaned up if disk space is low. On Linux this will end up in
either `$XDG_DATA_HOME` or `~/.local/store`. On Mac it'll end up in
`~/Library/Application Support`. On Windows it'll end up in
`~/AppData/Roaming`.

Because we're hitting an external API for something that is strictly an
optional feature, I've been very defensive about making sure we hit the
API *at most* once per day. We pessimistically update the mtime before
actually making the request, so that if something goes wrong we don't
end up spamming their API after every run. If an error does occur for
some reason, it will be silently ignored in release mode. In debug
builds it will panic if the error happened in the parent process, or
print to stderr if it happened in the child.

This uses `ureq`, which is definitely not a common or standard HTTP
client. I would normally go with `reqwest`, but that ends up pulling in
the entire tokio ecosystem, since it's async by default. I didn't want
to add that sort of impact on our compile times. `ureq` is a lighter
weight HTTP client that suits our needs, and is widely used enough that
I'm comfortable with depending on it.