Refactor argument handling in main (#302)
This moves all the argument handling from a series of hard-ish to follow
if statements to pattern matching, and also adds support for file paths
which are not valid UTF-8. Since we can't pattern match a specific value
for an OsStr nicely, I've done the slightly funky setup of matching on
the first argument converted to a string, and then the actual array.
This suits our needs for now, and if we need more than this we should
probably switch to a proper argument parser like clap anyway.