Remove ruby version from submodule path
As we update the version of Ruby over time we're not going to want to
update a bunch of hard coded version numbers. The `ruby_checkout/README`
file seemed redundant, as a git submodule is not the same as
redistributing code, so there shouldn't be any license or copyright
concerns. Because of that I've just moved the submodule up a level

Related to #315.