Support ruby 2.7 args forwarding syntax (#5)
* Add tests

* Implement support for argument forwarding in method definitions and bodies

* fmt

* Add test for multiline args forwarding in return

This adds support for the argument forwarding syntax added in Ruby 2.7, which is used like this:

```
def my_method(...)
  other_method(...)
end
```

This PR adds support for this as a method argument, although note that I'm not sure if this will properly handle the forwarding syntax used in 3.0, since in 3.0, arguments can be partially forwarded or given additional arguments (e.g. def foo(first_arg, ...) or forwarded_to_method(:foo, ...). However, this is at least sufficient to unblock the usages in pay-server.