Correctly format multiline method calls with blocks (#14)
* Add fixtures

* Fix blockvar in multiline method calls with blocks

* Prevent newlines above do blocks in method calls

This fixes two bugs that arose from the same fixture. First, multiline method calls with a block variable would add a newline between the last method argument and the closing parenthesis, which was caused by all do keywords being prefixed with a newline instead of accounting for do blocks in multiline methods. (In this case, I've opted to still prefix single-line methods that contain a do-block with a newline, since that was the preexisting functionality.

The second issue was that block arguments in this case were being incorrectly multilined due them being marked as starting one line before, so rubyfmt attempted to maintain what it thought was already multilined. The fix for this is just winding the line forward when we have multilined arguments so that the starting line for the block argument is correct.