Properly multiline user-multilined call chains (#382)
In general, `rubyfmt` respects (within limits) user choices about multilining expressions, but we don't currently do this for multiline expressions. This has a few poor side effects: (1) users often _want_ to multiline to make e.g. adding comments for specific sections of a call chain easier, and (2) there end up being some (IMO) awkward formatting of chains like the following:
```ruby
MyClass.call(
  [],
  ""
).do_some_other_stuff!
```
which is inconsistent with `rubyfmt`'s normal call chain styling.

This PR implements multilining for user-multilined expressions. It does this by checking the line location of the `DotTypeOrOp` items in the call chain and multilining if they're on multiple lines. This _might_ be a slightly imperfect if the user writes something like this

```ruby
# both format as `foo.bar.baz`
foo
.bar.baz

foo.bar.
baz
```

but IMO this is a reasonable implementation tradeoff. We _could_ alternatively check every expression's start/end, but this ends up requiring a _lot_ more trickery and is more prone to inconsistency.